rename: upload-revenue api 폴더링
