+++ src/pages/admin/upload-revenue/[month].page.tsx
@@ -9,7 +9,7 @@ import SectionLayout from "@/components/common/Layouts/SectionLayout";
 import Loading from "@/components/common/Loading/Loading";
 import MonthPickerDropdown from "@/components/common/MonthPicker/MonthPickerDropdown";
 import UploadHistroyTable from "@/components/upload-revenue/UploadHistoryTable/UploadHistoryTable";
-import { getRevenueUploadHistory } from "@/services/api/upload-revenue-mock-api";
+import { getRevenueUploadHistory } from "@/services/api/upload-revenue/upload-revenue-mock-api";
 import { useUploadHistoryGet } from "@/services/queries/upload-revenue/useRevenueUploadHistory";
 import { MEMBER_TYPE } from "@/types/enums/user.enum";
 import convertYearMonthToQuery from "@/utils/convertYearMonthToQuery";
